+++ src/index.ts
@@ -1,0 +1,3 @@
+const str: string = '1';
+
+document.write('Hello world' + str);